refactor: 전체적인 구조 변경
- utils 폴더에 적합하지 않은 파일 분리
- 메서드 순서 변경
- 메서드 및 변수 이름 변경
- 가독성을 위한 공백라인 추가
