+++ src/js/model/lottoPrices.js
